Remove apple dependencies, now builds and passes tests on linux
Still probably won't work on Windows due to SSL and SASL. There is no FindX for SASL, and
FindOpenSSL doesn't quite work the way we need, since it doesn't give you a library path, which
is what we need to send along to SCons. For now, it is assuming that the location where boost
comes from is sufficient to pick up the SSL and SASL libraries.